+++ flowmachine/flowmachine/core/server/query_schemas/reference_location.py
@@ -4,6 +4,9 @@
 
 from marshmallow_oneofschema import OneOfSchema
 
+from flowmachine.core.server.query_schemas.coalesced_location import (
+    CoalescedLocationSchema,
+)
 from flowmachine.core.server.query_schemas.daily_location import DailyLocationSchema
 from flowmachine.core.server.query_schemas.majority_location import (
     MajorityLocationSchema,
@@ -18,6 +21,10 @@ from flowmachine.core.server.query_schemas.visited_most_days import (
 
 
 class ReferenceLocationSchema(OneOfSchema):
+    """
+    A set of queries that return a mapping between unique subscribers and locations
+    """
+
     type_field = "query_kind"
     type_schemas = {
         "daily_location": DailyLocationSchema,
@@ -25,4 +32,5 @@ class ReferenceLocationSchema(OneOfSchema):
         "most_frequent_location": MostFrequentLocationSchema,
         "visited_most_days": VisitedMostDaysSchema,
         "majority_location": MajorityLocationSchema,
+        "coalesced_location": CoalescedLocationSchema,
     }